backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
APScheduler==3.10.4
twilio==8.10.0

